requirements.txt
python-telegram-bot[ext]==22.6
httpx==0.28.1
python-dotenv==1.0.1
